Add installer tests
